Win64 size_t changes applied to reaction class definitions.

+++ src/IrreversibleExchangeReaction.h
@@ -53,7 +53,7 @@ namespace mesmer
     virtual Molecule *get_excessReactant(void) const { return m_rct2; };
 
     // Return products
-    virtual int get_products(std::vector<Molecule *> &product) const
+    virtual size_t get_products(std::vector<Molecule *> &product) const
     {
       product.push_back(m_pdt1);
       if (m_pdt2) {
@@ -63,7 +63,7 @@ namespace mesmer
       return 1;
     };
 
-    virtual int get_reactants(std::vector<Molecule *> &reactants) const
+    virtual size_t get_reactants(std::vector<Molecule *> &reactants) const
     {
       reactants.push_back(m_rct1);
       reactants.push_back(m_rct2);
